+++ server/controllers/product_controller.js
@@ -1,34 +1,16 @@
 const modelProduct = require('../models/modelProducts');
 const productController = {};
 const path = require('path'); //Manejamos directorios
-const fs = require('fs'); //Manejador de archivos de node
-productController.createProduct = async (req, res, next) => {
-  try {
-    const Product = new modelProduct({
-      name: req.body.name,
-      description: req.body.description,
-      price: req.body.price,
-      imagePath: req.file.path,
-      imageName: req.file.filename,
-    });
-    const resultado = await Product.save();
-    res.json({
-      resultado: resultado,
-      status: 'Product save',
-    });
-  } catch (error) {
-    next(error);
-  }
-};
+const fs = require('fs').promises; //Manejador de archivos de node
+
 productController.getAllProducts = async (req, res, next) => {
   try {
     const productos = await modelProduct.find();
-    res.json(productos);
+    return res.json(productos);
   } catch (error) {
     next(error);
   }
 };
-
 productController.getProduct = async (req, res, next) => {
   try {
     const producto = await modelProduct.findById(req.params.id);
@@ -37,30 +19,57 @@ productController.getProduct = async (req, res, next) => {
     next(error);
   }
 };
-productController.updateProduct = async (req, res, next) => {
+
+productController.createProduct = async (req, res, next) => {
   try {
-    const { id } = req.params;
-    const product = {
+    const data = {
       name: req.body.name,
       description: req.body.description,
       price: req.body.price,
-      imagePath: req.file.path,
-      imageName: req.file.filename,
     };
-    console.log(req.file);
-    //BUSCAMOS PRIMERO EL PRODUCTO PARA ELIMINAR LA FOTO ANTERIOR Y LUEGO AGREGAR LA NUEVA
-    const productDeletePhoto = await modelProduct.findById(req.params.id);
-    if (productDeletePhoto) {
-      //Si existe el producto a eliminar
-      fs.unlink(path.resolve(productDeletePhoto.imagePath), (err) => {
-        //Eliminamos la foto en la ruta que tenia guardado el producto
-        if (err) throw err;
-        console.log('Successfully delete');
-      });
+    if (req.file !== undefined) {
+      data.imagePath = req.file.path;
+      data.imageName = req.file.filename;
+    }
+    const product = new modelProduct(data);
+    const resultado = await product.save();
+    return res.json({
+      resultado: resultado,
+      status: 'Product save',
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+productController.updateProduct = async (req, res, next) => {
+  try {
+    const { id } = req.params;
+    const data = {
+      name: req.body.name,
+      description: req.body.description,
+      price: req.body.price,
+    };
+    if (req.file !== undefined) {
+      data.imagePath = req.file.path;
+      data.imageName = req.file.filename;
+      //BUSCAMOS PRIMERO EL PRODUCTO PARA ELIMINAR LA FOTO ANTERIOR Y LUEGO AGREGAR LA NUEVA
+      const productDeletePhoto = await modelProduct.findById(req.params.id);
+      if (productDeletePhoto) {
+        //Si existe el producto a eliminar
+        if (productDeletePhoto.imageName !== 'no-image.svg') {
+          //EVITAMOS BORRAR LA IMAGEN ESTANDAR POR SI NO HAY IMAGENES
+          await fs.unlink(path.resolve(productDeletePhoto.imagePath), (err) => {
+            //Eliminamos la foto en la ruta que tenia guardado el producto
+            if (err) throw err;
+            console.log('Successfully delete');
+          });
+        }
+      }
     }
     const resultado = await modelProduct.findOneAndUpdate(
       id,
-      { $set: product },
+      { $set: data },
       { new: true }
     );
     return res.json({
@@ -74,26 +83,19 @@ productController.updateProduct = async (req, res, next) => {
 
 productController.deleteProduct = async (req, res, next) => {
   try {
-    const productPhoto = await modelProduct.findById(req.params.id);
-    //Si existe el producto a eliminar
-    if (productPhoto) {
-      console.log(path.resolve(productPhoto.imagePath));
-      if (path.resolve(productPhoto.imagePath)) {
-        fs.unlink(path.resolve(productPhoto.imagePath), (err) => {
-          //Eliminamos la foto en la ruta que tenia guardado el producto
-          if (err) {
-            throw err;
-          }
-          res.json({
-            status: 'File succesfully deleted',
-          });
-        });
+    const { id } = req.params;
+    const product = await modelProduct.findByIdAndRemove(id);
+    if (product) {
+      if (product.imageName !== 'no-image.svg') {
+        try {
+          await fs.unlink(path.resolve(product.imagePath));
+          console.log('File removed');
+        } catch (err) {
+          console.error('No pudimos eliminar la imagen', err);
+        }
       }
-      const product = await modelProduct.findByIdAndDelete(req.params.id);
     }
-    return res.json({
-      status: 'Producto eliminado',
-    });
+    return res.json({ status: 'Product Deleted' });
   } catch (error) {
     next(error);
   }
